Consolidate Vocab support code into one place
