+++ jena-fuseki2/jena-fuseki-access/src/main/java/org/apache/jena/fuseki/access/VocabSecurity.java
@@ -21,7 +21,7 @@ package org.apache.jena.fuseki.access;
 import org.apache.jena.rdf.model.Property;
 import org.apache.jena.rdf.model.Resource;
 import org.apache.jena.sparql.core.assembler.AssemblerUtils;
-import org.apache.jena.tdb.assembler.Vocab;
+import org.apache.jena.system.Vocab;
 
 public class VocabSecurity {
     private static final String NS = "http://jena.apache.org/access#";
