+++ jena-tdb1/src/main/java/org/apache/jena/tdb/assembler/VocabTDB.java
@@ -21,6 +21,7 @@ package org.apache.jena.tdb.assembler;
 import org.apache.jena.rdf.model.Property ;
 import org.apache.jena.rdf.model.Resource ;
 import org.apache.jena.sparql.core.assembler.AssemblerUtils ;
+import org.apache.jena.system.Vocab;
 import org.apache.jena.tdb.TDB ;
 
 public class VocabTDB
