+++ jena-text/src/main/java/org/apache/jena/query/text/assembler/TextVocab.java
@@ -20,7 +20,7 @@ package org.apache.jena.query.text.assembler;
 
 import org.apache.jena.rdf.model.Property ;
 import org.apache.jena.rdf.model.Resource ;
-import org.apache.jena.tdb.assembler.Vocab ;
+import org.apache.jena.system.Vocab ;
 
 public class TextVocab
 {
